move 'Preloader' component to 'src/components' folder
