+++ src/components/Preloader/Preloader.scss
