+++ src/modules/pages/NewApiPage/NewApiItem.js
@@ -8,7 +8,7 @@ import getUpdatedEndpoint from '../../../helpers/getUpdatedEndpoint';
 import Section from '../../Layout/Section/Section';
 import Subtitle from '../../Layout/Title/Subtitle';
 import EndpointForm from '../../forms/EndpointForm/EndpointForm';
-import Preloader from '../../Preloader/Preloader';
+import Preloader from '../../../components/Preloader/Preloader';
 
 const propTypes = {
     api: PropTypes.object.isRequired,
