+++ src/modules/pages/OAuthServersPage/OAuthServersList.js
@@ -11,7 +11,7 @@ import ROUTES from '../../../configurations/routes.config';
 
 import PaginatedList from '../../PaginatedList/PaginatedList';
 import Icon from '../../Icon/Icon';
-import Preloader from '../../Preloader/Preloader';
+import Preloader from '../../../components/Preloader/Preloader';
 import NoSearchResults from '../../../components/NoSearchResults/NoSearchResults';
 
 import '../../Layout/Table/Table.css';
